agents/requirements.txt
google-cloud-aiplatform>=1.66.0
google-auth
pygithub
pyyaml
